refactor: extract mail api handlers
